// src/types/document.d.ts
interface DocumentTypes {
    id: string
    folder_id: string
    create_user: string
    create_time: string
    operate_time: string
    operate_user: string
    request_data: string
    response_data: string
    project_id: number
    description: string
    name: string
}

export type { DocumentTypes }